requirements.txt
Django==1.7.1
Pillow
-e git+https://github.com/mariocesar/sorl-thumbnail.git#egg=sorl-thumbnail
django-reversion

